fix: correct invalid UINT32 SQL type in audit_record migration
audit_record.i2c_address was declared as UINT32, which is not a valid
PostgreSQL type (SQLSTATE 42704 on migration apply). Corrected to
SMALLINT, matching the convention already used for sensor.i2c_address
and sensor_hw_history.i2c_address (migration 013).

Discovered when branches merging this migration (via #1199, #1193)
failed to apply schema in integration tests.

Part of #1166

+++ leaflab/migrate/migrations/022_audit_record.up.sql
@@ -19,7 +19,7 @@ CREATE TABLE audit_record (
     reason          TEXT,                   -- optional explanation (e.g., skip reason)
     -- Additional columns for config application context (supporting FR1.3):
     config_version  BIGINT,                 -- device_config.version being applied
-    i2c_address     UINT32,                 -- hardware address of the affected sensor (if applicable)
+    i2c_address     SMALLINT,               -- hardware address of the affected sensor (if applicable)
     mux_path        JSONB                   -- mux path of the affected sensor (if applicable)
 );
